Add files via upload
Hello! This is a new version that I am currently working on! I looking to complete a larger amount of work tonight including finishing animations and uploading the animals into the project. I will also see if I can upload the color of the animals back in once that is completed. If you have any suggestions or ideas about how I should integrate some ideas, you can send them through GitHub or Gmail. I'm excited to continue the work on the project and see it transform!