Add files via upload
Here is the first set of files that I am uploading here. I thought that I would try uploading all the files in my project folder to see if that will allow you to open the project properly! I have a few more to add in another batch, but I hope this may work better!